stro from unsigned to double.

+++ include/living.h
@@ -8,7 +8,7 @@ struct Living
     unsigned age;
     int stamina;
     unsigned vit;   /* The HP has to be proportional to its VIT. */
-    unsigned str;   /* The attack has to be proportional to its STR. */
+    double str;   /* The attack has to be proportional to its STR. */
     CommonAtrrs attrs;
     unsigned weight;
 };
